organize the sessions into their own grouped files
