+++ .github/workflows/pr_checks.yml
@@ -57,13 +57,13 @@ jobs:
         run: pip install nox>=2022
 
       - name: Check fidesctl installation
-        run: nox -s check-install -- docker
+        run: nox -s check_install -- docker
 
       - name: Run fidesctl evaluation
         run: nox -s fidesctl -- docker
 
       - name: Scan fidesctl db
-        run: nox -s fidesctl-db-scan -- docker
+        run: nox -s fidesctl_db_scan -- docker
 
   isort:
     needs: Build
